fix(ring-071): Rename token.py to auth_token.py to avoid stdlib conflict
